requirements.txt
picamera==1.10
requests==2.9.1
RPi.GPIO==0.6.1
wheel==0.24.0
